basic curd operation - Grahql

+++ src/main/java/com/graphQl/learn/repository/BookRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepo extends JpaRepository<Book,Long> {
+public interface BookRepo extends JpaRepository<Book,Integer> {
 }
